third_party/blink/web_tests/platform/linux-chrome/external/wpt/webrtc-encoded-transform/tentative/RTCPeerConnection-insertable-streams-video.https-expected.txt

This is a testharness.js-based test.
[FAIL] Frames flow correctly using insertable streams
  promise_test: Unhandled rejection with value: object "InvalidStateError: Failed to execute 'createEncodedStreams' on 'RTCRtpSender': Encoded video streams not requested at PC initialization"
[FAIL] Frames flow correctly using insertable streams when receiver starts negotiation
  promise_test: Unhandled rejection with value: object "InvalidStateError: Failed to execute 'createEncodedStreams' on 'RTCRtpSender': Encoded video streams not requested at PC initialization"
[FAIL] Creating streams twice throws
  promise_test: Unhandled rejection with value: object "InvalidStateError: Failed to execute 'createEncodedStreams' on 'RTCRtpSender': Encoded video streams not requested at PC initialization"
[FAIL] Encoded frames serialize and deserialize into a deep clone
  promise_test: Unhandled rejection with value: object "InvalidStateError: Failed to execute 'createEncodedStreams' on 'RTCRtpSender': Encoded video streams not requested at PC initialization"
Harness: the test ran to completion.
